[New] added testing of the *byId() methods
Here i found a conceptual problem where the *byId() methods didn't
properly resolve ObjectRefs and add them to the TX because, we added the
ObjectRefs to the filter, which made the TX resolve PersistenContexts by
the wrong type.

Now we always add the same type of object to the ObjectFilter; the
PersistentContext. This means that when we call ObjectDao.add(T), etc.
we must already create such the context. This helps resolve another
issue at the back of my mind: We don't want to partially commit a TX if
such a basic information as the object's CTX is not available.