[New] added state to the container - checked by PrivilegeHandler
The container how sets its own state once it has changed the state of
the components. Now the DefaultStrolchPrivilegeHandler checks that the
container has started before allowing any methods to be called. 

This might still be too restrictive in that when the container is
stopped, that then an admin can not log in and perform any actions, but
this can then be changed when needed.