[Major] Strolch agent is now fully realm capable
ResourceMap and OrderMap are kept in a StrolchRealm object for each
realm which is configured. This is transparent for the user. When no
realms are defined then the 'defaultRealm' is configured where all
methods without a realm parameter use this realm.

Once a realm is configured, the defaultRealm must be configured if it 
is still to be used at runtime.

Tests make sure realms work, for both transient, and empty situations.
Cached agent mode is next to be implemented