[Major] Major rewrite of Controller/Activity execution
Controller now never triggers execution on the ExecutionHandler, but calls itself as often as necessary to keep execution of the same Activity in the same thread run. This leads to a more predictable sequence of executing of actions.

Further it is now possible to override the Controller's instantiated, to hook into the trigger of further execution of an Activity.